pyproject.toml
[tool.poetry]
name = "near-ai"
version = "0.1.0"
description = "Helper AI tools for Near"
authors = ["Reza Rahemtola <reza.rahemtola@epitech.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
